Refactor Armstrong number functions to support Int128 and update test cases

+++ exercises/practice/armstrong-numbers/.meta/Sources/ArmstrongNumbers/ArmstrongNumbersExample.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-func isArmstrongNumber(_ number: Int) -> Bool {
+func isArmstrongNumber(_ number: Int128) -> Bool {
     let stringNumber = String(number)
     let numberLength = stringNumber.count
     let sum = stringNumber.compactMap { Int(String($0)) }.reduce(0) { $0 + Int(pow(Double($1), Double(numberLength))) }
