+++ exercises/practice/armstrong-numbers/Sources/ArmstrongNumbers/ArmstrongNumbers.swift
@@ -1,3 +1,3 @@
-func isArmstrongNumber(_ number: Int) -> Bool {
+func isArmstrongNumber(_ number: Int128) -> Bool {
     // Write your code for the 'Armstrong Numbers' exercise here.
 }
